+++ pygccxml/declarations/calldef.py
@@ -18,7 +18,6 @@ This modules contains definition for next C++ declarations:
 """
 from . import cpptypes
 from . import declaration
-from . import class_declaration
 from . import calldef_types
 
 
@@ -307,9 +306,13 @@ class calldef_t(declaration.declaration_t):
         self._has_inline = has_inline
 
     def _report(self, *args, **keywd):
-        return class_declaration.dependency_info_t(self, *args, **keywd)
+        # Implementation detail. Will be removed when the deprecated
+        # i_depend_on_them method is dropped
+        from . import dependencies  # pylint: disable=R0401
+        return dependencies.dependency_info_t(self, *args, **keywd)
 
     def i_depend_on_them(self, recursive=True):
+        self._warn_deprecated()
         answer = []
         if self.return_type:
             answer.append(