+++ pygccxml/declarations/class_declaration.py
@@ -16,9 +16,9 @@ from . import scopedef
 from . import declaration_utils
 from . import declaration
 from . import templates
-from . import cpptypes
 from . import byte_info
 from . import elaborated_info
+from . import type_traits
 
 
 class ACCESS_TYPES(object):
@@ -165,6 +165,7 @@ class class_declaration_t(declaration.declaration_t):
         return []
 
     def i_depend_on_them(self, recursive=True):
+        self._warn_deprecated()
         return []
 
     @property
@@ -460,12 +461,16 @@ class class_t(
         return answer
 
     def i_depend_on_them(self, recursive=True):
+        self._warn_deprecated()
+        # Deprecated method. The cyclic import will be removed with the method
+        # in the next release, so we can disable the cyclic import check here.
+        from . import dependencies  # pylint: disable=R0401
 
         answer = []
 
         for base in self.bases:
             answer.append(
-                dependency_info_t(
+                dependencies.dependency_info_t(
                     self,
                     base.related_class,
                     base.access_type,
@@ -478,7 +483,6 @@ class class_t(
         return answer
 
     def _get_partial_name_impl(self):
-        from . import type_traits  # prevent cyclic dependencies
         if type_traits.is_std_string(self):
             return 'string'
         if type_traits.is_std_wstring(self):
@@ -501,99 +505,3 @@ class class_t(
 
 
 class_types = (class_t, class_declaration_t)
-
-
-class impl_details(object):
-
-    @staticmethod
-    def dig_declarations(depend_on_it):
-
-        # FIXME: prevent cyclic imports
-        from . import type_traits
-
-        if isinstance(depend_on_it, declaration.declaration_t):
-            return [depend_on_it]
-        base_type = type_traits.base_type(
-            type_traits.remove_alias(depend_on_it))
-        if isinstance(base_type, cpptypes.declarated_t):
-            return [base_type.declaration]
-        elif isinstance(base_type, cpptypes.calldef_type_t):
-            result = []
-            result.extend(impl_details.dig_declarations(base_type.return_type))
-            for argtype in base_type.arguments_types:
-                result.extend(impl_details.dig_declarations(argtype))
-            if isinstance(base_type, cpptypes.member_function_type_t):
-                result.extend(
-                    impl_details.dig_declarations(
-                        base_type.class_inst))
-            return result
-        return []
-
-
-class dependency_info_t(object):
-
-    def __init__(self, decl, depend_on_it, access_type=None, hint=None):
-        object.__init__(self)
-
-        assert isinstance(depend_on_it, (class_t, cpptypes.type_t))
-        self._decl = decl
-        self._depend_on_it = depend_on_it
-        self._access_type = access_type
-        self._hint = hint
-
-    @property
-    def declaration(self):
-        return self._decl
-
-    @property
-    def depend_on_it(self):
-        return self._depend_on_it
-
-    @property
-    def access_type(self):
-        return self._access_type
-
-    @access_type.setter
-    def access_type(self, access_type):
-        self._access_type = access_type
-
-    def __str__(self):
-        return 'declaration "%s" depends( %s ) on "%s" ' \
-               % (self.declaration, self.access_type, self.depend_on_it)
-
-    @property
-    def hint(self):
-        """The declaration, that report dependency can put some additional
-        inforamtion about dependency. It can be used later"""
-        return self._hint
-
-    def find_out_depend_on_it_declarations(self):
-        """If declaration depends on other declaration and not on some type
-        this function will return reference to it. Otherwise None will be
-        returned
-        """
-        return impl_details.dig_declarations(self.depend_on_it)
-
-    @staticmethod
-    def i_depend_on_them(decl):
-        """Returns set of declarations. every item in the returned set,
-        depends on a declaration from the input"""
-
-        to_be_included = set()
-        for dependency_info in decl.i_depend_on_them():
-            for ddecl in dependency_info.find_out_depend_on_it_declarations():
-                if ddecl:
-                    to_be_included.add(ddecl)
-
-        if isinstance(decl.parent, class_t):
-            to_be_included.add(decl.parent)
-        return to_be_included
-
-    @staticmethod
-    def we_depend_on_them(decls):
-        """Returns set of declarations. every item in the returned set,
-        depends on a declaration from the input"""
-        to_be_included = set()
-        for decl in decls:
-            to_be_included.update(dependency_info_t.i_depend_on_them(decl))
-        return to_be_included